+++ TodoWithNoRequestDelegate/TodoApi.cs
@@ -25,7 +25,7 @@ namespace Todos
 
         private async Task GetAsync(HttpContext context)
         {
-            if (!context.Request.RouteValues.TryGet("id", out long id))
+            if (!context.Request.RouteValues.TryGet("id", out int id))
             {
                 context.Response.StatusCode = 400;
                 return;
@@ -50,7 +50,7 @@ namespace Todos
 
         private async Task DeleteAsync(HttpContext context)
         {
-            if (!context.Request.RouteValues.TryGet("id", out long id))
+            if (!context.Request.RouteValues.TryGet("id", out int id))
             {
                 context.Response.StatusCode = 400;
                 return;
